Store strings not latents; add prompt_idx argument
This gives a stronger guarantee that I'm not messing with existing code.
At first I thought something was wrong with the first prompt, but now I
realize the issue is I checked "if prompt_idx" instead of "if prompt_idx
is not None". The former doesn't trigger if prompt_idx is 0.