Debug seed interpolation
It doesn't work with pure img2img, but it does work when I use a mask.
This puts us in the inpainting model instead of img2img. However the
code looks basically identical in the problematic parts, so I think the
real issue may be deeper, in the sampler. At any rate, for now I'm just
always going to use a mask.